Refactored AdvancementOptions into separate component and started building styling for overlay
