+++ package.json
@@ -16,7 +16,7 @@
     "react-force-graph-2d": "^1.6.1",
     "react-ga": "^2.5.3",
     "react-overlays": "^0.8.3",
-    "react-redux": "^5.1.1",
+    "react-redux": "^5.1.2",
     "react-reorder": "^3.0.0-alpha.7",
     "react-router-dom": "^4.3.1",
     "react-scripts": "^3.3.0",
