+++ src/components/FloatingButton/FloatingButton.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import './floatingButton.css'
+
+const FloatingButton = () => {
+    return (
+        <div className="floatingButton mobileOnly">
+            <button>
+                <span>+</span>
+            </button>
+        </div>
+    );
+}
+export default FloatingButton;
