+++ src/components/MonsterFinder/MonsterFinder.css
@@ -17,8 +17,5 @@
   .flex-container { flex-wrap: wrap; }
   .flex-item:first-child { flex-basis: 100%; }
   .flex-item:nth-child(2) { text-align:left; }
-  .flexSelect {
-    border: 1px solid gray;
-    width: calc(100% - 2px);
-  }
+
 }
