+++ src/components/MonsterFinder/MonsterFinder.js
@@ -1,42 +1,28 @@
 import React, { Component } from 'react';
 
 import './MonsterFinder.css';
-//import 'react-select/dist/react-select.css';
 import '../../css/ReactSelectCustom.css';
+import './subcomponents/advancementOptions_mobile.css';
 
 import { connect } from 'react-redux'
-import {
-  keyPressHandler, 
-  monsterSelectedHandler, hitDiceAdvancementAction, sizeAdvancementAction, abilityScoreAdvancementAction, 
-  templateAdvancementAction, classLevelAdvancementAction} from '../../action-creators'
+import {keyPressHandler, monsterSelectedHandler} from '../../action-creators'
 import PathfinderMonsterAdvancer from '../PathfinderMonsterAdvancer/PathfinderMonsterAdvancer'
 import MonsterOptions from '../MonsterOptions'
 import MonsterSelect from './MonsterSelect'
 import Aasimar from '../../models/AasimarV2'
 
 import PageViewRecorder from '../../components/PageViewRecorder';
-import HitDiceAdvancementSelectMaterial from './subcomponents/HitDiceAdvancementSelectMaterial';
-import SizeAdvancementSelectMaterial from './subcomponents/SizeAdvancementSelectMaterial';
-import AbilityScoreAdvancementSelectMaterial from './subcomponents/AbilityScoreAdvancementSelectMaterial';
 
-import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
 import MonsterDisplay from '../MonsterDisplay';
-import { MonsterSizes } from '../PathfinderMonsterAdvancer/AdvancementTools/MonsterSizes';
-import TemplateSelect from './subcomponents/TemplateSelect';
-import { template } from '@babel/core';
-import ClassLevelSelect from '../CleverSelect/ClassLevelSelect';
+import FloatingButton from '../FloatingButton/FloatingButton';
+import AdvancementOptions from './subcomponents/AdvancementOptions';
 
 export class MonsterFinder extends Component {
   constructor(props) {
     super(props);
     this.handleKeyPress = this.handleKeyPress.bind(this);
     this.handleMonsterSelectChange = this.handleMonsterSelectChange.bind(this);
-    this.handleHitDiceSelectChange = this.handleHitDiceSelectChange.bind(this);
-    this.handleSizeSelectChange = this.handleSizeSelectChange.bind(this);
-    this.handleAbilityScoreSelectChange = this.handleAbilityScoreSelectChange.bind(this);
-    this.handleTemplateChange = this.handleTemplateChange.bind(this);
-    this.classLevelsChanged = this.classLevelsChanged.bind(this);
-    this.classLevelSelectRef = React.createRef();
+    this.advancementOptionsRef = React.createRef();
   }
 
   handleKeyPress(e) {
@@ -45,38 +31,9 @@ export class MonsterFinder extends Component {
 
   handleMonsterSelectChange(e, {suggestion}) {
     this.props.monsterSelectedHandler(suggestion);
-    this.props.hitDiceAdvancementAction('reset');
-    this.props.sizeAdvancementAction('reset');
-    this.props.abilityScoreAdvancementAction('resetall');
-    this.props.templateAdvancementAction('reset');
-    this.props.classLevelAdvancementAction('reset');
-    this.classLevelSelectRef.current.reset();
+    console.log(this.advancementOptionsRef.current)
+    this.advancementOptionsRef.current.getWrappedInstance().reset();
   }
-
-  handleHitDiceSelectChange(e) {
-    this.props.hitDiceAdvancementAction(parseInt(e.target.value));
-  }
-
-  handleSizeSelectChange(e) {
-    this.props.sizeAdvancementAction(e.target.value);
-  } 
-
-  handleTemplateChange(selectedValues) {
-    this.props.templateAdvancementAction(selectedValues);
-  }
-
-  handleAbilityScoreSelectChange(e, abilityScore) {
-    //where abilityScore is the identifier like 'Str'
-    //console.log(e.target.value, abilityScore);
-    this.props.abilityScoreAdvancementAction(parseInt(e.target.value), abilityScore);
-  }
-
-  classLevelsChanged(classLevels) {
-    //console.log("MonsterFinder.classLevelsChanged", classLevels)
-    //if (!classLevels || classLevels.length <= 0) return;
-    //const mappedClassLevels = classLevels.map(x => x.className+x.level)
-    this.props.classLevelAdvancementAction(classLevels);
-}
 
   componentDidMount() {
     document.addEventListener('keydown', this.handleKeyPress);
@@ -86,30 +43,8 @@ export class MonsterFinder extends Component {
     if (this.props.match && this.props.match.params && this.props.match.params.monsterName) {
       this.props.monsterSelectedHandler(this.props.match.params.monsterName);
     }
-    if (this.props.location.search) {
-      const searchParams = new URLSearchParams(this.props.location.search);
-      this.props.abilityScoreAdvancementAction(parseInt(searchParams.get("str")), 'str');
-      this.props.abilityScoreAdvancementAction(parseInt(searchParams.get("dex")), 'dex');
-      this.props.abilityScoreAdvancementAction(parseInt(searchParams.get("con")), 'con');
-      this.props.abilityScoreAdvancementAction(parseInt(searchParams.get("int")), 'int');
-      this.props.abilityScoreAdvancementAction(parseInt(searchParams.get("wis")), 'wis');
-      this.props.abilityScoreAdvancementAction(parseInt(searchParams.get("cha")), 'cha');
-
-      const sizeToParse = searchParams.get("size");
-      if (sizeToParse) {
-        const size = MonsterSizes.find(x => x.abbr === sizeToParse.toUpperCase() || x.size.toLowerCase() === sizeToParse.toLowerCase());
-        if (size)
-          this.props.sizeAdvancementAction(size.size);
-      }
-      this.props.hitDiceAdvancementAction(parseInt(searchParams.get("hd")));
-      const templatesToParse = searchParams.get('templates');
-      if (templatesToParse) {
-        this.props.templateAdvancementAction(templatesToParse.split(","));
-      }
-    }
 
     PageViewRecorder.recordPageView(window.location.pathname + window.location.search, undefined, title);
-
   }
 
   componentWillUnmount() {
@@ -121,50 +56,33 @@ export class MonsterFinder extends Component {
   //TODO: Skills and knowledges getting updated by stats lose extras (knowledge loses subtype, any specific bonuses are also lost)
   render() {
     let { monster, advancement } = this.props; //These props can be destructured to pull out any of the reducers (config, select, monster, s3Select, etc.)
-    const theme = createMuiTheme({
-      overrides: {
-        MuiOutlinedInput: {
-          input: {
-            padding: 8,
-          }
-        }
-      },
-      typography: {
-        useNextVariants: true,
-      },
-    });
+    
     monster = (monster.success) ? monster : { success: true, statBlock: Aasimar};
     const advancedMonster = PathfinderMonsterAdvancer(monster, advancement)
-    const templatesOption = true;
-    const templateSelect = (templatesOption) ? <TemplateSelect selectedTemplates={advancement.templates} onSelect={this.handleTemplateChange}/> : '';
+    
     return (
-      <MuiThemeProvider theme={theme}>
-      <div className="flex-container">
-        <div className="flex-item">
-          <MonsterDisplay monster={advancedMonster}/>
-        </div>
-        <div className="flex-item">
-          <div className="flexSelect" style={{backgroundColor: 'white'}}>
-            <MonsterSelect listItems={MonsterOptions.map(op => op.props.children)} onSelect={this.handleMonsterSelectChange}/>
-            <HitDiceAdvancementSelectMaterial originalHitDice={monster.statBlock.hitDice} selectedHitDice={advancement.hd} onSelect={this.handleHitDiceSelectChange}/>
-            <SizeAdvancementSelectMaterial originalSize={monster.statBlock.size} selectedSize={advancement.size} onSelect={this.handleSizeSelectChange} />
-            <AbilityScoreAdvancementSelectMaterial selectedValue={advancement.str} abilityScore={"Str"} onSelect={this.handleAbilityScoreSelectChange}/>
-            <AbilityScoreAdvancementSelectMaterial selectedValue={advancement.dex} abilityScore={"Dex"} onSelect={this.handleAbilityScoreSelectChange}/>
-            <AbilityScoreAdvancementSelectMaterial selectedValue={advancement.con} abilityScore={"Con"} onSelect={this.handleAbilityScoreSelectChange}/>
-            <AbilityScoreAdvancementSelectMaterial selectedValue={advancement.int} abilityScore={"Int"} onSelect={this.handleAbilityScoreSelectChange}/>
-            <AbilityScoreAdvancementSelectMaterial selectedValue={advancement.wis} abilityScore={"Wis"} onSelect={this.handleAbilityScoreSelectChange}/>
-            <AbilityScoreAdvancementSelectMaterial selectedValue={advancement.cha} abilityScore={"Cha"} onSelect={this.handleAbilityScoreSelectChange}/>
-            {templateSelect}
-            <ClassLevelSelect hideLabel classes={["Barbarian", "Bard", "Cleric"]} onChange={(e) => this.classLevelsChanged(e)} ref={this.classLevelSelectRef}/>
+      <React.Fragment>
+        <div className="flex-container">
+          <div className="flex-item">
+            <MonsterDisplay monster={advancedMonster}/>
+          </div>
+          <div className="flex-item">
+            <div className="flexSelect" style={{backgroundColor: 'white'}}>
+              <MonsterSelect listItems={MonsterOptions.map(op => op.props.children)} onSelect={this.handleMonsterSelectChange}/>
+              <AdvancementOptions {...this.props} ref={this.advancementOptionsRef}/>
+            </div>
           </div>
         </div>
-    </div>
-    </MuiThemeProvider>
+        <div>
+          <FloatingButton/>
+        </div>
+        <div className="advancement_overlay">
+          <AdvancementOptions {...this.props}/>
+        </div>
+      </React.Fragment>
     );
   }
 }
 
 const mapStateToProps = state => state;
-
-
-export default connect(mapStateToProps, {keyPressHandler, monsterSelectedHandler, hitDiceAdvancementAction, sizeAdvancementAction, abilityScoreAdvancementAction, templateAdvancementAction, classLevelAdvancementAction})(MonsterFinder)
+export default connect(mapStateToProps, {keyPressHandler, monsterSelectedHandler})(MonsterFinder)
